fix: narrow react-native exports
