+++ .changeset/react-native-storage-exports.md
@@ -1,0 +1,5 @@
+---
+"accounts": patch
+---
+
+Make `accounts/react-native` adapter-only and move React Native storage adapters behind explicit subpaths.
