+++ src/react-native/asyncStorage.ts
@@ -2,37 +2,6 @@ import { Json } from 'ox'
 
 import * as Storage from '../core/Storage.js'
 
-/** Creates a storage adapter backed by `expo-secure-store`. */
-export function secureStorage(options: secureStorage.Options = {}): Storage.Storage {
-  return Storage.from(
-    {
-      async getItem(name) {
-        const { getItemAsync } = await import('expo-secure-store')
-        const raw = await getItemAsync(name)
-        if (raw === null) return null
-        try {
-          return Json.parse(raw)
-        } catch {
-          return null
-        }
-      },
-      async setItem(name, value) {
-        const { setItemAsync } = await import('expo-secure-store')
-        await setItemAsync(name, Json.stringify(value))
-      },
-      async removeItem(name) {
-        const { deleteItemAsync } = await import('expo-secure-store')
-        await deleteItemAsync(name)
-      },
-    },
-    options,
-  )
-}
-
-export declare namespace secureStorage {
-  type Options = Storage.from.Options
-}
-
 /** Creates a storage adapter backed by `@react-native-async-storage/async-storage`. */
 export function asyncStorage(options: asyncStorage.Options = {}): Storage.Storage {
   return Storage.from(
@@ -61,5 +30,6 @@ export function asyncStorage(options: asyncStorage.Options = {}): Storage.Storag
 }
 
 export declare namespace asyncStorage {
+  /** Options for `asyncStorage`. */
   type Options = Storage.from.Options
 }
