+++ src/react-native/index.ts
@@ -1,3 +1,1 @@
-export * as Provider from './Provider.js'
 export { reactNative } from './adapter.js'
-export { asyncStorage, secureStorage } from './storage.js'
